Extract test helpers to token_partition_tree_test_utils
