+++ common/formatting/align_test.cc
@@ -22,6 +22,7 @@
 #include "absl/strings/str_split.h"
 #include "common/formatting/format_token.h"
 #include "common/formatting/token_partition_tree.h"
+#include "common/formatting/token_partition_tree_test_utils.h"
 #include "common/formatting/unwrapped_line_test_utils.h"
 #include "common/text/tree_builder_test_util.h"
 #include "common/util/range.h"
@@ -1434,94 +1435,6 @@ TEST(ColumnsTreeFormatter, ColumnPositionTreePrinter) {
   }
 }
 
-// TODO(mglb): copied from layout_optimizer_test.cc. Extract it to separate file
-// and reuse.
-class TokenPartitionTreeBuilder {
- public:
-  TokenPartitionTreeBuilder(
-      int indent, std::pair<int, int> token_indexes_range,
-      PartitionPolicyEnum policy,
-      std::initializer_list<TokenPartitionTreeBuilder> children = {})
-      : indent_(indent),
-        token_indexes_range_(token_indexes_range),
-        policy_(policy),
-        children_(children) {}
-
-  // TODO(mglb): this constructor is new
-  TokenPartitionTreeBuilder(
-      int indent, PartitionPolicyEnum policy,
-      std::initializer_list<TokenPartitionTreeBuilder> children = {})
-      : indent_(indent), policy_(policy), children_(children) {}
-
-  TokenPartitionTreeBuilder(
-      int indent, std::pair<int, int> token_indexes_range,
-      std::initializer_list<TokenPartitionTreeBuilder> children = {})
-      : indent_(indent),
-        token_indexes_range_(token_indexes_range),
-        children_(children) {}
-
-  TokenPartitionTreeBuilder(
-      std::pair<int, int> token_indexes_range, PartitionPolicyEnum policy,
-      std::initializer_list<TokenPartitionTreeBuilder> children = {})
-      : token_indexes_range_(token_indexes_range),
-        policy_(policy),
-        children_(children) {}
-
-  TokenPartitionTreeBuilder(
-      std::pair<int, int> token_indexes_range,
-      std::initializer_list<TokenPartitionTreeBuilder> children = {})
-      : token_indexes_range_(token_indexes_range), children_(children) {}
-
-  TokenPartitionTreeBuilder(
-      PartitionPolicyEnum policy,
-      std::initializer_list<TokenPartitionTreeBuilder> children)
-      : policy_(policy), children_(children) {}
-
-  TokenPartitionTreeBuilder(
-      std::initializer_list<TokenPartitionTreeBuilder> children)
-      : children_(children) {}
-
-  TokenPartitionTree build(
-      const std::vector<verible::PreFormatToken>& tokens) const {
-    TokenPartitionTree node;
-
-    auto& child_nodes = node.Children();
-    child_nodes.reserve(children_.size());
-    for (const auto& child : children_) {
-      node.NewChild(child.build(tokens));
-    }
-
-    FormatTokenRange node_tokens;
-    if (token_indexes_range_.first < 0) {
-      CHECK(!child_nodes.empty());
-      CHECK_LT(token_indexes_range_.second, 0);
-      node_tokens.set_begin(child_nodes.front().Value().TokensRange().begin());
-      node_tokens.set_end(child_nodes.back().Value().TokensRange().end());
-    } else {
-      CHECK_GE(token_indexes_range_.second, token_indexes_range_.first);
-      node_tokens.set_begin(tokens.begin() + token_indexes_range_.first);
-      node_tokens.set_end(tokens.begin() + token_indexes_range_.second);
-    }
-
-    node.Value() = UnwrappedLine(indent_, node_tokens.begin(), policy_);
-    node.Value().SpanUpToToken(node_tokens.end());
-    return node;
-  }
-
- private:
-  int indent_ = 0;
-  std::pair<int, int> token_indexes_range_ = {-1, -1};
-  PartitionPolicyEnum policy_ = PartitionPolicyEnum::kUninitialized;
-
-  const std::vector<TokenPartitionTreeBuilder> children_;
-};
-
-bool PartitionsEqual(const UnwrappedLine& left, const UnwrappedLine& right) {
-  return (left.TokensRange() == right.TokensRange()) &&
-         (left.IndentationSpaces() == right.IndentationSpaces()) &&
-         (left.PartitionPolicy() == right.PartitionPolicy());
-}
-
 // Delimiter that matches text outside of substrings between 'start' and 'stop'
 // (inclusive).
 class OutsideCharPairs {
@@ -1574,22 +1487,6 @@ class FormatUsingOriginalSpacingTest : public ::testing::Test,
   }
 
  protected:
-  static testing::AssertionResult TokenPartitionTreesEqualPredFormat(
-      const char* actual_expr, const char* expected_expr,
-      const TokenPartitionTree& actual, const TokenPartitionTree& expected) {
-    const auto diff = DeepEqual(actual, expected, PartitionsEqual);
-    if (diff.left != nullptr) {
-      return testing::AssertionFailure()
-             << "Expected equality of these trees:\n"
-                "Actual:\n"
-             << actual
-             << "\n"
-                "Expected:\n"
-             << expected << "\n";
-    }
-    return testing::AssertionSuccess();
-  }
-
   void RunTestCase(TokenPartitionTree actual,
                    const TokenPartitionTree expected) {
     std::vector<TokenPartitionTree> nodes;
